requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
google-genai
chromadb
sentence-transformers
langchain-text-splitters
streamlit
python-dotenv
httpx
requests
pymupdf
python-docx
